requirements.txt
python-dotenv
requests
PyQt6
